[feat] Add comments for the dashboard types
