+++ aim/web/ui/src/pages/Home/components/ExperimentsCard/ExperimentsCard.scss
@@ -2,7 +2,7 @@
 
 .ExperimentsCard {
   border-top: $border-main;
-  padding: 1rem;
+  padding: $space-unit;
   &__title {
     margin-bottom: $space-unit;
   }
@@ -18,6 +18,7 @@
     flex-wrap: wrap;
     .SearchInput {
       margin-bottom: $space-xs;
+      margin-right: 0;
     }
   }
   .selectedSelectIcon {
